+++ Core/Registry/Registry.cs
@@ -1001,9 +1001,8 @@ namespace CKAN
                 log.DebugFormat("Started with {0}, removing {1}, and keeping {2}; our dlls are {3}", string.Join(", ", orig_installed), string.Join(", ", modules_to_remove), string.Join(", ", hypothetical), string.Join(", ", dlls));
 
                 // Find what would break with this configuration.
-                // The Values.SelectMany() flattens our list of broken mods.
-                var broken = new HashSet<string>(SanityChecker.FindUnmetDependencies(hypothetical, dlls)
-                    .Values.SelectMany(x => x).Select(x => x.identifier));
+                var broken = SanityChecker.FindUnsatisfiedDepends(hypothetical, dlls.ToHashSet())
+                    .Select(x => x.Key.identifier).ToHashSet();
 
                 // If nothing else would break, it's just the list of modules we're removing.
                 HashSet<string> to_remove = new HashSet<string>(modules_to_remove);